feat: Add getArch and getPlatform ipc handlers
