+++ packages/preload/src/ipc.ts
@@ -34,8 +34,3 @@ export function launchExplorer(version?: string) {
 export function launchState(cb: (event: IpcRendererEvent, state: IpcRendererEventData) => void) {
   return ipcRenderer.on(IPC_EVENTS.LAUNCH_EXPLORER, cb);
 }
-
-export async function getOSName(): Promise<string> {
-  const resp = await ipcRenderer.invoke(IPC_HANDLERS.GET_OS_NAME);
-  return resp;
-}
